chore: rename QRCodeV2 to QR_Code and remove useMemo hook from the useStyles
