+++ packages/design-system/src/atoms/index.tsx
@@ -23,4 +23,4 @@ export * from './tooltipV2'
 export * from './statusV2'
 export * from './switch'
 export * from './radio-select'
-export * from './qrcodeV2'
+export * from './qrcode'
