+++ packages/design-system/src/atoms/qrcode/qrcode.stories.tsx
@@ -1,18 +1,18 @@
 import { titleBuilder, fromTemplate } from '../../../.storybook/utils'
-import { QRCodeProps, QRCodeV2 } from '.'
+import { QRCodeProps, QR_Code } from '.'
 
 const DEFAULT_VALUE = 'https://tryroll.com/'
 const DEFAULT_SIZE = 200
 
 const conf = {
-  title: titleBuilder.atoms('QRCodeV2'),
-  component: QRCodeV2,
+  title: titleBuilder.atoms('QRCode'),
+  component: QR_Code,
 }
 
 const Template = ({
   value = DEFAULT_VALUE,
   size = DEFAULT_SIZE,
-}: QRCodeProps) => <QRCodeV2 value={value} size={size} />
+}: QRCodeProps) => <QR_Code value={value} size={size} />
 
 export const Default = fromTemplate(Template, {
   value: DEFAULT_VALUE,
